Initial Code Commit
conpiguator.js uses vanilla javascript. So, it has no dependencies, but it is not IE compatible.

conpigurator.jquery.js is dependent on jQuery methods to become more succinct and provide IE compatibility. If you are already loading jQuery, this file has an even smaller footprint. If IE compatibility is necessary, jQuery + this file will have you conpigurating in a flash.